price changed to double

+++ src/main/java/org/example/wishlist/Repository/WishlistRepository.java
@@ -81,7 +81,7 @@ public class WishlistRepository {
                 rs.getInt("wishListID"),
                 rs.getString("name"),
                 rs.getString("description"),
-                rs.getInt("price"),
+                rs.getDouble("price"),
                 rs.getString("link")));
     }
 
@@ -93,7 +93,7 @@ public class WishlistRepository {
                 rs.getInt("wishListID"),
                 rs.getString("name"),
                 rs.getString("description"),
-                rs.getInt("price"),
+                rs.getDouble("price"),
                 rs.getString("link")));
     }
     public Wishes getWishByWishID(int ID) {
@@ -103,7 +103,7 @@ public class WishlistRepository {
                 rs.getInt("wishListID"),
                 rs.getString("name"),
                 rs.getString("description"),
-                rs.getInt("price"),
+                rs.getDouble("price"),
                 rs.getString("link")));
     }
